tweaks to ccp4 import

+++ src/ccp4.h
@@ -7,25 +7,25 @@
 
 struct CCP4_MAP_HEADER
 {
-  long nc,nr,ns;
-  long mode;
-  long ncstart,nrstart,nsstart;
-  long nx,ny,nz;
+  int nc,nr,ns;
+  int mode;
+  int ncstart,nrstart,nsstart;
+  int nx,ny,nz;
   float x,y,z;
   float alpha,beta,gamma;
-  long mapc,mapr,maps;
-  long amin,amax,amean;
-  long ispg;
-  long nsymbt;
-  long lskflag;
-  long skwmat[9];
-  long skwtrn[3];
-  long dummy[15];
-  long map;
-  long machst;
+  int mapc,mapr,maps;
+  int amin,amax,amean;
+  int ispg;
+  int nsymbt;
+  int lskflag;
+  int skwmat[9];
+  int skwtrn[3];
+  int dummy[15];
+  int map;
+  int machst;
   float arms;
-  long nlabl;
-  long label[200];
+  int nlabl;
+  int label[200];
 };
 
 int ccp4Read(FILE *f, dbmScalNode *node);
